RangeSelectorも分離
