+++ column/range_test.go
